+++ dispatch/fixtures_test.go
@@ -1,4 +1,4 @@
-package cmd
+package dispatch
 
 var (
 	getWorkflowsResponse string = `{
